requirements.txt
# Core dependencies (pinned to avoid resolver backtracking in Docker)
numpy==1.26.0
pandas==2.1.4
scikit-learn==1.3.2

# Configuration
pyyaml==6.0.1

# Metrics and visualization
matplotlib==3.8.2
seaborn==0.13.0

# Progress bars
tqdm==4.65.0

# Embedding generation dependencies
biopython==1.81
transformers==4.35.2
tokenizers==0.14.1
huggingface-hub==0.17.3
sentencepiece==0.2.1
h5py==3.10.0
einops==0.7.0
attrs==23.2.0
cloudpathlib==0.23.0
httpx==0.28.1
tenacity==9.1.2
biotite==1.2.0
brotli==1.2.0
msgpack==1.1.2
msgpack-numpy==0.4.8
zstd==1.5.7.3

# ESM-C SDK is installed in Dockerfile (avoid heavy resolver backtracking)
